requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Faker==15.1.1
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
